Move generateLinkFileDefault from the shared function before. Remove the obsolete generateLinkFile
