+++ src/generate/generateAppFiles/generatorDefault/index.ts
@@ -1,5 +1,7 @@
 import { default as generateUseRedirectFile } from './generateUseRedirectFileDefault';
+import { default as generateLinkFile } from './generateLinkFileDefault';
 
 export default {
   generateUseRedirectFile,
+  generateLinkFile,
 };
